requirements.txt
openai>=1.40.0
python-dotenv>=1.0.0
streamlit>=1.40.0
pypdf>=5.2.0
